pyproject.toml
[project]
name = "ja-dubbing"
version = "7.0.0"
description = "英語動画を話者音声クローン付き日本語吹き替え動画に変換"
requires-python = ">=3.13"
dependencies = [
    "pip",
    "python-dotenv>=1.0.0",
    "spacy>=3.7.0",
    "pydub>=0.25.0",
    "audioop-lts>=0.2.1",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "numpy>=1.26.0",
    "pyannote-audio>=4.0.4",
    "llama-cpp-python>=0.3.0",
    "huggingface-hub>=0.20.0",
    "sentencepiece>=0.2.1",
    "mlx>=0.25.0",
    "mlx-lm>=0.20.0",
    "mlx-audio[stt]>=0.3.0",
    "soundfile>=0.12.0",
    "omnivoice>=0.1.2",
    "transformers>=5.3.0",
    "accelerate>=0.20.0",
    "langdetect>=1.0.0",
    "torchcodec==0.7.0",
]


[project.scripts]
ja-dubbing = "ja_dubbing.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ja_dubbing"]

[tool.uv]
override-dependencies = [
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "torchcodec==0.7.0",
    "transformers>=5.3.0",
]

[tool.ruff]
line-length = 99
target-version = "py313"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["ja_dubbing"]
